Update TicTacToe.cpp
After two pull requests, I found out that the newest pull request overwrote a previous pull request. The reason I believe is that the newest commit represents the newest version of the file, so the pull request of the newest version will delete the older changes if they are on the same line of code.